telegram-bot/requirements.txt
python-telegram-bot==20.0a1
Pillow==9.2.0
aiohttp==3.8.1
python-dotenv==1.0.1
